% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rule_building.R
\name{add_cumulative}
\alias{add_cumulative}
\title{Add cumulative numbers of positive and negative matches to a list of rules}
\usage{
add_cumulative(data, order_by = "score", rule_var = "rule")
}
\arguments{
\item{data}{A Rules data frame.}

\item{order_by}{A column in \code{data} according to which order rules in
descending order.}

\item{rule_var}{The name of the column in \code{data} which stores the rules.}
}
\value{
The Rules data frame with two more columns for the cumulative numbers
of positive and negative matches.
}
\description{
Given a Rules data frame, arranges it by a score and rule length and computes
the cumulative number of positive and negative matches identified by the
rules.
}
