% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rule_building.R
\name{add_negative_terms}
\alias{add_negative_terms}
\title{Add negative terms to rules}
\usage{
add_negative_terms(rules, target_vec, target_data)
}
\arguments{
\item{rules}{A vector of rules as produced by \code{\link[=extract_rules]{extract_rules()}}.}

\item{target_vec}{A vector of labels.}

\item{target_data}{A DTM with a number of rows as the elements in
\code{rules}.}
}
\value{
A data frame with a column with the negative terms for each
rule, the number of positive and negative matches identified by the
negative terms in the \code{target_data} and the ID of these records..

pos/neg' = glue('{tot_pos}/{tot_neg}'),
}
\description{
Increases rule specificity by adding negative terms iteratively, selecting
only terms that remove non-relevant records without impacting the number of
positive matches. Usually not used by itself, but inside
\code{\link[=generate_rule_selection_set]{generate_rule_selection_set()}}.
}
