% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{analyse_grid_search}
\alias{analyse_grid_search}
\title{Analyse the results of a parameter grid search}
\usage{
analyse_grid_search(
  session_folder = "Grid_Search",
  tot_pos = NULL,
  tot_records = NULL,
  plot = TRUE,
  score = c("Sens_adj_eff", "Pos_rate", "Pos_rate_adj_sens")
)
}
\arguments{
\item{session_folder}{Where to find the result sessions produced by
\code{\link{perform_grid_evaluation()}}.}

\item{tot_pos}{Total number of positive matches among records. If \code{NULL}
it will be inferred by the Annotation files in \code{session_folder} which
then need to be fully labelled.}

\item{tot_records}{Total number of records. If \code{NULL} it will be
inferred by the Annotation files in \code{session_folder} which then need
to be fully labelled.}

\item{plot}{Whether to plot the marginal impact of each parameter.}

\item{score}{Which one of the scores to use to measure classification
performance. Can be \code{Sens_adj_eff}: sensitivity by efficacy,
\code{Pos_rate}: the ratio of positive labels found over the total record
(positive rate), \code{Pos_rate_adj_sens}: sensitivity by positive rate.}
}
\value{
A list with: \item{iterations}{A data frame describing each
classification/review iteration for each session, reporting the parameter
values used and the performance score.}\item{best_parms}{The highest
performance parameter set in the best parameter cluster. First, the
parameter clusters are ordered by average performance score, then parameter
combinations inside the cluster are ordered by sensitivity followed by
efficiency. A data frame with the score, the ratio of positive matches over
the total records, the ratio of reviewed records, the sensitivity and the
efficiency, is reported.}\item{best_by_rule}{A data frame with performance
info (like for \code{best_parms}) for the best parameter set for each
performance cluster.}\item{plot}{A plot with the marginal impact of each
parameter if the \code{plot} argument is \code{TRUE}, otherwise
\code{NULL}.}
}
\description{
Takes as input a folder with multiple session data produced by
\code{\link{perform_grid_evaluation}()}, each session representing a
combination of parameters and computes the best combinations in terms of a
perfomance score.
}
\details{
For each session a performance score is computed. The default is
\eqn{\text{sensitivity} \times \text{efficiency}}{sensitivity x efficiency},
with \eqn{efficiency} being the \emph{one minus the ratio of reviewed records
over the total}. The statistics are computed on a subset of fully labeled
records.

The analysis is performed on each session's "Results" files.

A partition tree algorithm is used to group parameter combinations by average
scores, identifying "performance clusters". The combination with the best
sensitivity followed by the best efficiency is then shown for each cluster.
Optionally, a plot can be generated with the marginal impact of each
parameter.
}
\examples{
\dontrun{
out <- analyse_grid_search()

# check the best parameter set (use str() for easy to read output)
str(out$best_parms)

# check the best set for each performance cluster
View(out$best_by_rule)

# plot parameter marginal impact is the plot argument is TRUE
out$plot
}

}
