% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_search.R
\name{clean_date_filter_arg}
\alias{clean_date_filter_arg}
\title{Transform a year range into the format required by a specific search engine}
\usage{
clean_date_filter_arg(
  year_query,
  cases,
  arg_in_query_test = NULL,
  query = NULL
)
}
\arguments{
\item{year_query}{A string identifying a search period (see Details).}

\item{cases}{The search engine syntax specification into which convert the
year query It needs to be a list with the following items: gt (greater
than), ge (greater than or equal to), lt (lower than), le (lower than or
equal to), eq (equal to), range (range of years). Inside the specification,
the year in the query can be retrieved with the \code{year_piece}
placeholder, which in the case of a closed range is a vector of two
elements (see examples).}

\item{arg_in_query_test}{A search engine syntax tag that identifies a year
filter. Useful to define an already existing year filter in the query. If
\code{arg_in_query_test} is set, also the \code{query} argument must be
set.}

\item{query}{The search query. Needed to test if a year filter is already
present. If \code{query} is set, also the \code{arg_in_query_test} argument
must be set.}
}
\value{
The formatted, search engine specific, period filter.
}
\description{
Take a period definition and transform it into a specific format passed
through the \code{cases} argument.
}
\details{
The input period specified in \code{year_query} must follow this format:
\itemize{\item{\code{> Year}: }{greater than \code{Year}.}\item{\code{>=
Year}: }{greater then or equal to \code{Year}.}\item{\code{< Year}: }{lower
than \code{Year}.}\item{\code{<= Year}: }{lower then or equal to
\code{Year}.}\item{\code{Year}: }{equal to
\code{Year}.}\item{\code{Year1-Year2}: }{range from \code{Year1} to
\code{Year2}.}}

The input is checked to evaluate if \code{year_query} the format is clearly
invalid.
}
\examples{

query <- "Test query"
year_query <- "2010-2020"

library(lubridate)

## Example for Web Of Science
year_query <- clean_date_filter_arg(year_query, cases = list(
gt = '{year_piece + 1}-{year(today())}', ge = '{year_piece}-{year(today())}',
eq = '{year_piece}-{year_piece}', le = '1985-{year_piece}',
range = '{year_piece[1]}-{year_piece[2]}', lt = '1985-{year_piece - 1}'),
arg_in_query_test = 'PY ?=', query = query)

## Example for Pubmed
year_query <- clean_date_filter_arg(year_query, cases = list(
gt = '{year_piece + 1}[PDAT]:{year(today())}[PDAT]',
ge = '{year_piece}[PDAT]:{year(today())}[PDAT]',
eq = '{year_piece}[PDAT]:{year_piece}[PDAT]',
le = '1000[PDAT]:{year_piece}[PDAT]',
lt = '1000[PDAT]:{year_piece - 1}[PDAT]',
range = '{year_piece[1]}[PDAT]:{year_piece[2]}[PDAT]'),
arg_in_query_test = '[PDAT]', query = query)


}
