% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{clean_record_textfields}
\alias{clean_record_textfields}
\title{Clean up problematic text in the citation data.}
\usage{
clean_record_textfields(df)
}
\arguments{
\item{df}{The data frame to be cleaned.}
}
\value{
The cleaned data frame.
}
\description{
Used internally by the \code{parse_*} functions.
}
