% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{coalesce_labels}
\alias{coalesce_labels}
\title{Join classification labels into one}
\usage{
coalesce_labels(
  data,
  label_cols = c("Rev_prediction_new", "Rev_prediction", "Rev_manual")
)
}
\arguments{
\item{data}{An Annotation data frame.}

\item{label_cols}{The label columns to coalesce. They are considered
hierarchically from left to right, with a lower level column being used if
a value is missing in the upper level one.}
}
\value{
A vector with the coalesce labels.
}
\description{
In the same Annotation file, there could be multiple classification labels
for a record derived by different phases of the labelling procedure. This
function joins them hierarchically, using the label from the lower
hierarchies to fill up \code{NA}s in the upper ones. The hierarchy is
considered from left to right.
}
\details{
This function is a wrapper over
\code{\link[dplyr:coalesce]{dplyr::coalesce}()} which allow to have a
standard coalescing scheme for an annotation file.
}
