% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{compute_BART_model}
\alias{compute_BART_model}
\title{Use Bayesian Additive Regression Trees to predict records labels}
\usage{
compute_BART_model(
  train_data,
  Y,
  preds = NULL,
  save = FALSE,
  folder = getwd(),
  name = as.character(Y),
  rebuild = FALSE,
  num_trees = 50,
  k = 2,
  num_iterations_after_burn_in = 2000,
  run_in_sample = FALSE,
  mem_cache_for_speed = TRUE,
  use_missing_data = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{train_data}{A Document Term Matrix with an outcome column. if
\code{pred} is null, all column apart from \code{Y} will be considered as
features.}

\item{Y}{The name of the outcome column in the data set.}

\item{preds}{A vector of column names to use as features.}

\item{save}{Whether to save the model to disk, with a file name given in
\code{name}, to be placed into \code{folder}.}

\item{folder}{Where to look for/create the model backup file.}

\item{name}{The name of the model backup file. If the file exists and
\code{rebuilt} is \code{FALSE}, the model will not be retrained but the
backup will be used instead.}

\item{rebuild}{If \code{TRUE}, retrain the model even if a model backup file
exists.}

\item{...}{More argument to pass to
\code{\link[bartMachine:bartMachine]{bartMachine::bartMachine}()}}

\item{um_trees, k, num_iterations_after_burn_in, run_in_sample, mem_cache_for_speed, use_missing_data, verbose}{\code{\link[bartMachine:bartMachine]{bartMachine::bartMachine}()} specific
parameters.}
}
\value{
An object of class \code{bartMachine}.
}
\description{
The prediction engine of the framework. It produces a model which
assigns a probability distribution for each record of the probability of
being relevant (i.e., positive label). It is not used alone but inside
\code\link{enrich_annotation_file}.
}
\details{
This implementation is built over
\code{\link[bartMachine:bartMachine]{bartMachine::bartMachine}}.
}
