% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{compute_changes}
\alias{compute_changes}
\title{Describe changes in the record labels after a Classification/Review Iteration}
\usage{
compute_changes(Annotations)
}
\arguments{
\item{Annotations}{An Annotation data set, as data frame or as a file path to
it.}
}
\value{
A data frame with total reviewed positive and negative labels and
their sum, the new labels added in the \code{Rev_prediction_new} column,
and and summary of all changes in the last iteration, e.g.: \code{n -> y}
(a previously non relevant record becoming relevant) or \code{unlab. -> n}
(an unlabeled record becoming negative).
}
\description{
Takes an Annotation file with a \code{Rev_prediction_new} column and computes
differences in the labelling introduced by the last automatic classification
/ review.
}
\examples{
\dontrun{

# Taking the path to an annotation data file; also the actual data frame can
# be passed
Annotations <- get_session_files("Session1")$Annotations[2]

out <- compute_changes(Annotations)

# The output is easier to read if changed in a long format:
str(out) # the easiest solution for fast inspection of results

# To get a long format data frame
tidyr::pivot_longer(out, everything(),
  names_to = "Indicator", values_to = "Value")
}
}
