% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{consolidate_results}
\alias{consolidate_results}
\title{Include manual review in annotation result summaries}
\usage{
consolidate_results(
  session_name,
  sessions_folder = getOption("baysren.sessions_folder")
)
}
\arguments{
\item{session_name}{The session with the result files to consolidate.}

\item{sessions_folder}{The path to the folder where all the sessions are
stored.}
}
\value{
The list of the regenerated Results' file content as data frames,
invisibly.
}
\description{
\code{\link{enrich_annotation_file}()} creates summary files of each
classification iteration into the Results session sub-folder, but these
summaries do not includes the manual review of the predicted labels.
\link{consolidate_results} regenerates these files including the changes
resulting from the manual review of the automatic classification.
}
\details{
The results summary, not including the manual review, is still available
inside the annotation files.
}
