% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Model.R
\name{create_training_set}
\alias{create_training_set}
\title{Create the training set for the automatic classification}
\usage{
create_training_set(Records, min_freq = 0.05)
}
\arguments{
\item{Records}{An Annotation data frame.}

\item{min_freq}{Minimum document frequency (between 0 and 1) in negative
labelled records above which a term is considered.}
}
\value{
An enriched DTM which is the merge of the title, abstract, authors,
keywords and MESH terms DTMs plus the record ID and label if present.
}
\description{
Given an Annotation file, it creates and join Document Term Matrices (DTM)
for the title, abstract, authors, keywords and MESH terms of a record,
keeping terms above a given document frequency among the negative labelled
records and which are present in at least two records
}
\examples{
\dontrun{

Records <- import_data(get_session_files("Session1")$Records)

DMT <- create_training_set(Records)
}
}
