% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_embedding_cache.R
\name{.embedding_text_roles}
\alias{.embedding_text_roles}
\title{Supported embedding text roles}
\usage{
.embedding_text_roles()
}
\value{
Character vector of supported text roles.
}
\description{
Supported embedding text roles
}
\keyword{internal}
