% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_embedding_cache.R
\name{.empty_embedding_cache}
\alias{.empty_embedding_cache}
\title{Create an empty embedding cache table}
\usage{
.empty_embedding_cache()
}
\value{
A zero-row embedding cache tibble.
}
\description{
Create an empty embedding cache table
}
\keyword{internal}
