% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_embedding_cache.R
\name{.parse_gemini_embedding_response}
\alias{.parse_gemini_embedding_response}
\title{Parse a Gemini embedding response}
\usage{
.parse_gemini_embedding_response(response)
}
\arguments{
\item{response}{An \code{httr2} response.}
}
\value{
A numeric embedding matrix.
}
\description{
Parse a Gemini embedding response
}
\keyword{internal}
