% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{DTM.aggr_synonyms}
\alias{DTM.aggr_synonyms}
\title{Aggregate redundant terms}
\usage{
DTM.aggr_synonyms(DTM, min.sim = 0.9)
}
\arguments{
\item{DTM}{A Document Term Matrix.}

\item{min.sim}{The minimal cosine similarity that identifies an edge.}
}
\value{
The same input Document Term Matrix with redundant terms removed and
joined into new columns.
}
\description{
Build a term-term network using a cosine similarity measure built on the term
co-presence in documents. A high threshold defined in \code{min.sim} is used
to identify edges. The high edge threshold splits the network into multiple
components which identify redundant terms.
}
