% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analysis.R
\name{estimate_positivity_rate_model}
\alias{estimate_positivity_rate_model}
\title{Train a simple Bayesian logistic model on classification results}
\usage{
estimate_positivity_rate_model(train_data, seed = 14129189)
}
\arguments{
\item{train_data}{An Annotation data set with predictions produced by
\code{\link{enrich_annotation_file}()}.}

\item{seed}{An integer to replicate results}
}
\value{
An object of class \code{brmsfit}. See
\code{\link[brms:brm]{brms::brm}()} for mode info.
}
\description{
The function uses an Annotation data set to train a Bayesian logistic model
that estimates the probability of a relevant record given the lower
boundaries of the PPD produced by the classification model for the records
whose label was manually reviewed.
}
\details{
The produced model can be used to predict the distribution of the cumulative
number of total unseen positive records.

Usually this function is not to be used directly but through
\code{\link{estimate_performance}()}.
}
