% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rule_building.R
\name{extract_rules}
\alias{extract_rules}
\title{Extract screening rules from an Annotation data set}
\usage{
extract_rules(
  session_name,
  rebuild_dtm = FALSE,
  vimp.threshold = 1.25,
  n.trees = 800,
  sessions_folder = getOption("baysren.sessions_folder"),
  save_path = file.path(sessions_folder, session_name, "rule_data.rds"),
  ...
)
}
\arguments{
\item{session_name}{A session identifier corresponding to folders into the
\code{sessions_folder} folder.}

\item{rebuild_dtm}{Whether to use the last DTM stored in the
\code{session_name} folder (\code{FALSE}) or rebuild it from the last
Annotation file (\code{TRUE}).}

\item{vimp.threshold}{A threshold in the standardized variable importance
score to filter out less relevant terms in the DTM.}

\item{n.trees}{How many draws to use from the PPD matrix to build decision
trees. This parameter strongly impacts computational time but increases
sensitivity of the rules found.}

\item{sessions_folder}{Where to find the \code{sessions} folders.}

\item{save_path}{Since generating the rules is a computation intense process
it's advisable to save the output in a .rds file placed inside the
\code{session_name} folder. User need only to provide the name of the file.}

\item{...}{Additional arguments passed to
\code{\link[rpart:rpart.control]{rpart::rpart.control()}}.}
}
\value{
A list with: \item{SpecificDTM}{The DTM with the less relevant terms
being filtered out and terms in multiple record fields being
aggregated.}\item{DTM}{The full DTM with the predicted
classification.}\item{rules}{A data frame reporting the selected rules with
the average PPD.}
}
\description{
Starting from a Document Term Matrix (DTM) and a posterior predictive
distribution (PPD) matrix produced by the Bayesian classification engine, a
decision tree algorithm is used to extract rules that partition a subset of
draws from the PPD. Beware that the generation of the rules may take a long
time.
}
\details{
The algorithm allows to use only a subset of the terms in the DTM and of the
samples in the PPD matrix to cut on computation time. In the first case, a
threshold is used to filter only the most relevant features in the DTM.
Before being used, terms in the DTM are aggregated if they appear in multiple
fields of the citation records and only their general presence in the record
will be stored.
}
