% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{extract_source_file_paths}
\alias{extract_source_file_paths}
\title{Extract the path to citation records files}
\usage{
extract_source_file_paths(
  journal,
  sessions = journal$Session_ID,
  queries = journal$Query_ID,
  sources = journal$Source,
  records_folder = "Records"
)
}
\arguments{
\item{journal}{A data frame produced by \code{\link{perform_search_session}}
or a file path to it.}

\item{sessions, queries, sources}{Sessions, queries and sources for the which
one wants to get the record data. By default, all record file paths are
retrieved.}

\item{records_folder}{The path to the folder where the records are stored. By
default is named \code{Records}.}
}
\value{
A vector of file paths.
}
\description{
The information on the records' location is stored in the session journal
created by \code{\link{perform_search_session}}. It is possible to select
records from specific session, query, or source combinations. Only parsed or
API downloaded record paths will be returned, not the raw data source files.
}
\examples{

\dontrun{
journal <- perform_search_session(
  query = query, year_query = year_filter,
  session_name = 'Session1', query_name = 'Query1',
  records_folder = 'Records',
  journal = 'Session_journal.csv')

  # using the journal as data frame
  paths <- extract_source_file_paths(journal)

  # using the journal path
  paths <- extract_source_file_paths('Session_journal.csv')
}
}
