% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{fix_duplicated_records}
\alias{fix_duplicated_records}
\title{Resolve duplicated records in a record data frame}
\usage{
fix_duplicated_records(records)
}
\arguments{
\item{record_list}{A list of data frames as created by
\code{\link{read_bib_files}()}.}
}
\value{
A record data frame with no duplicated records.
}
\description{
This function uses the DOI or the (cleaned) record title to match duplicated
records and join them. Some fields like the record keywords and source are
joined between copies.
}
