% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting.R
\name{format_interval}
\alias{format_interval}
\title{Format a 3-values statistic}
\usage{
format_interval(interval, percent = FALSE)
}
\arguments{
\item{interval}{A 3 values vector describing a point estimate and two interval
boundaries.}

\item{percent}{Whether to format the results as percentages.}
}
\value{
A string in the "point statistic \link{interval boundaries}" format.
}
\description{
Useful to format a 3-values statistic in the "point statistic [interval
boundaries]" format.
}
\examples{

format_interval(qbeta(c(.05, .5, .95), 10, 14), percent = TRUE)

}
