% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting.R
\name{format_performance}
\alias{format_performance}
\title{Pretty formatting of Session performance analysis}
\usage{
format_performance(..., session_names = NULL)
}
\arguments{
\item{...}{One or more performance result data frames produced by
\code{\link{estimate_performance}()}.}

\item{session_names}{Names of the sessions corresponding to the result data
frames passed to \code{...}. If missing, they will be "Session" followed by
an incremental number for each data frame passed to \code{...}.}
}
\value{
A long format data frame with the statistical indicators on the first
column and a column with values for each data frame passed to \code{...}.
}
\description{
Publication friendly formatting of the results of
\code{\link{estimate_performance}()}. If more than one results set is passed
(i.e., one per session), the will be added as new columns.
}
\examples{

\dontrun{
Performance <- list(
 s1 = get_session_files('Session1')$Annotations \%>\% last() \%>\%
  import_data() \%>\% estimate_performance(),
 s2 = get_session_files('Session2')$Annotations \%>\% last() \%>\%
  import_data() \%>\% estimate_performance()
)

format_performance(Performance$s1, Performance$s2)
}
}
