% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting.R
\name{format_var_imp}
\alias{format_var_imp}
\title{Format variable importance results}
\usage{
format_var_imp(var_imp, as_data_frame = TRUE)
}
\arguments{
\item{var_imp}{A data frame produced by \code{\link{extract_var_imp}()}.}

\item{as_data_frame}{Whether to format the output as data frame or as text.}
}
\value{
A formatted data frame or a string of text, depending on the
\code{as_data_frame} argument.
}
\description{
Publication ready formatting of the output of
\code{\link[=extract_var_imp]{extract_var_imp()}}. Separate the term from the part of the
record it was found in; uses "&" and "|" to identify non-consecutive n-grams
and redundant terms; reduce numeric values to significant digits.
}
\examples{

\dontrun{
output <- extract_var_imp('Session1')

format_var_imp(output)
}
}
