% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting.R
\name{get_source_distribution}
\alias{get_source_distribution}
\title{Distribution of the number of sources in common per record}
\usage{
get_source_distribution(annotation_file, as_propr = TRUE, format_fun = percent)
}
\arguments{
\item{annotation_file}{An annotation data frame or a file path to it.}

\item{as_propr}{Whether to output the results as absolute numbers or
proportions over the record total.}

\item{format_fun}{If the results are shown as proportion (\code{as_propr ==
TRUE}), which function to use to format them.}
}
\value{
A named vector with the number of sources in common as names and the
absolute number/proportion as values.
}
\description{
Describe the distribution of the number of sources shared by records.
}
\examples{
\dontrun{
# Get the annotation file
Annotation_file <- get_session_files("Session1")$Annotations[1]

# Transform it into a list
get_source_distribution(Annotation_file)
}
}
