% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{get_website_resources}
\alias{get_website_resources}
\title{Tool to grab XHR messages from dynamic websites}
\usage{
get_website_resources(
  url,
  url_filter = ".*",
  type_filter = ".*",
  wait_for = 20,
  n_of_resources = NULL,
  interactive = FALSE
)
}
\description{
Tool to grab XHR messages from dynamic websites
}
