% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{import_classification}
\alias{import_classification}
\title{Import classifications from a previously labelled annotation data frame}
\usage{
import_classification(records, prev_records, IDs = records$ID)
}
\arguments{
\item{records}{An annotation data frame.}

\item{prev_records}{An annotation data frame with already existing manual or
automatic classification.}

\item{IDs}{Import the labels only from specific records.}
}
\value{
An annotation data frame with an extra column \code{Rev_previous}
storing the imported classification.
}
\description{
The classification is imported in a new column called \code{Rev_previous}.
}
