% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{join_records}
\alias{join_records}
\title{Join citation data frames and resolve record duplication}
\usage{
join_records(record_list)
}
\arguments{
\item{record_list}{A list of data frames as created by \code{\link[=read_bib_files]{read_bib_files()}}.}
}
\value{
A record data frame containing the citation data from multiple
sources.
}
\description{
Take a list of data frames containing parsed citation data created using
\code{\link[=read_bib_files]{read_bib_files()}} and joins them, resolving duplicated records using
\code{\link[=fix_duplicated_records]{fix_duplicated_records()}}.
}
\examples{
\dontrun{
data_files <- list.files(file.path("Records", "Session 1", "Query"))

record_list <- read_bib_files(data_files)

join_records(record_list)
}
}
