% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{lemmatize}
\alias{lemmatize}
\title{Transform words into their base form}
\usage{
lemmatize(text_vec, dict = lexicon::hash_lemmas)
}
\arguments{
\item{text_vec}{A vector of text documents.}

\item{dict}{A dictionary to use to find base forms. See
\code{\link[lexicon:hash_lemmas]{lexicon::hash_lemmas}()} for the required
structure.}
}
\value{
The vector with terms transformed into their base form, when
possible.
}
\description{
Transform words into their base form
}
