% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{order_by_query_match}
\alias{order_by_query_match}
\title{Reorder a data frame of records according to simple query match}
\usage{
order_by_query_match(records, query)
}
\arguments{
\item{records}{The data frame of records to reorder.}

\item{query}{The query to use to arrange the records.}
}
\value{
The ordered record data frame.
}
\description{
This function uses a simple algorithm that counts the keywords matches in
the query in the record abstract and title, and use the count to arrange the
records in descending order. Usually, the same query used to collect the
records in the first place are used.
}
