% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{parse_ieee}
\alias{parse_ieee}
\title{Parse IEEE raw data}
\usage{
parse_ieee(entries, timestamp = now())
}
\arguments{
\item{entries}{An imported data frame.}

\item{timestamp}{A timestamp as provided by
\code{\link[lubridate:now]{lubridate::now()}}.}
}
\value{
A data frame with the parsed data.
}
\description{
Parse and normalize IEEE data downloaded manually (as csv or excel files)
from \url{https://ieeexplore.ieee.org/Xplore/home.jsp}.
}
\examples{

\dontrun{
dataRaw <- import_data(file.path("Records", "Session 1", 'Query',
  'IEEE.csv'))

parse_ieee(dataRaw)
}
}
