% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{parse_pubmed}
\alias{parse_pubmed}
\title{Parse Pubmed raw data}
\usage{
parse_pubmed(entries, timestamp = now())
}
\arguments{
\item{entries}{A character vector containing the citation data.}

\item{timestamp}{A timestamp as provided by
\code{\link[lubridate:now]{lubridate::now}()}.}
}
\value{
A data frame with the parsed data.
}
\description{
Parse and normalize Pubmed ".nbib" files downloaded through
\url{https://pubmed.ncbi.nlm.nih.gov/}.
}
\examples{

\dontrun{
dataRaw <- readr::read_file(file.path("Records", "Session 1", 'Query',
  'Pubmed.nbib'))

parse_pubmed(dataRaw)
}
}
