% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{parse_scopus}
\alias{parse_scopus}
\title{Parse SCOPUS raw data}
\usage{
parse_scopus(entries, timestamp = now())
}
\arguments{
\item{entries}{An imported data frame.}

\item{timestamp}{A timestamp as provided by
\code{\link[lubridate:now]{lubridate::now()}}.}
}
\value{
A data frame with the parsed data.
}
\description{
Parse and normalize SCOPUS data downloaded manually (as csv or excel files)
from \url{https://www.scopus.com/search/form.uri?display=basic#basic}.
}
\examples{

\dontrun{
dataRaw <- import_data(file.path("Records", "Session 1", 'Query', 'EMBASE.csv'))

parse_embase(dataRaw)
}
}
