% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{parse_wos}
\alias{parse_wos}
\title{Parse Web of Science raw data}
\usage{
parse_wos(entries, timestamp = now())
}
\arguments{
\item{entries}{An imported data frame.}

\item{timestamp}{A timestamp as provided by
\code{\link[lubridate:now]{lubridate::now()}}.}
}
\value{
A data frame with the parsed data.
}
\description{
Parse and normalize Web of Science data downloaded manually (as CSV or Excel
files) from \url{https://www.webofknowledge.com}.
}
\examples{

\dontrun{
dataRaw <- import_data(file.path("Records", "Session 1", 'Query', 'WOS.csv'))

parse_wos(dataRaw)
}
}
