% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting.R
\name{print_table}
\alias{print_table}
\title{Publication friendly tables for .rmd files}
\usage{
print_table(data, caption = "", allow_math = FALSE, ...)
}
\arguments{
\item{data}{A data frame}

\item{caption}{A caption to be displayed in the table.}

\item{allow_math}{Whether to allow latex math by disabling special character
escape.}

\item{...}{Other arguments passed to
\code{\link[knitr:kable]{knitr::kable()}}}
}
\value{
An \code{\link[rkmarkdown:render]{rkmarkdown::render()}} ready table.
}
\description{
A publication friendly version of \code{\link[knitr:kable]{knitr::kable()}}.
It automatically detect if the output is HTML or PDF and adapt the
formatting, allowing for latex formulas, large table, etc in PDF outputs.
Allows using \\% in PDF tables.
}
