% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Record_management.R
\name{read_bib_files}
\alias{read_bib_files}
\title{Import and parse citation data files}
\usage{
read_bib_files(files)
}
\arguments{
\item{files}{A vector of citation data file paths.}
}
\value{
A list of data frames containing the parsed citation data.
}
\description{
Parse and normalize citation data files. If the file was already parsed, will
be just read.
}
\examples{

\dontrun{
data_files <- list.files(file.path("Records", "Session 1", 'Query'))

read_bib_files(data_files)
}
}
