% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rule_building.R
\name{simplify_ruleset}
\alias{simplify_ruleset}
\title{Remove redundant rules and rule components}
\usage{
simplify_ruleset(ruleset, target_vec, target_data)
}
\arguments{
\item{ruleset}{A rule set as generated by \code{\link[=generate_rule_selection_set]{generate_rule_selection_set()}}.}

\item{target_vec}{A vector of labels.}

\item{target_data}{A DTM with a number of rows as the elements in
\code{rules}.}
}
\value{
A data frame with the preserved rules, the redundant terms removed by
the rules, the number of positive and negative (absolute and cumulative)
records identified by the rules in the \code{target_data} and the position
of these records in the \code{target_data}.
}
\description{
Removes whole rules or components of the rules whose removal do not decrease
the total number of positive matches found by applying the rules in the
original data set.
}
\details{
The suggested approach is apply this function on the output of
\code{\link[=generate_rule_selection_set]{generate_rule_selection_set()}} after manual review of the proposed rules.
}
\examples{
\dontrun{
candidate_queries <- readRDS(file.path("Sessions", "Session1", "rule_data.rds"))

Target <- candidate_queries$DTM$Target
SpecificDTM <- candidate_queries$SpecificDTM

simplified_rules <- file.path("Sessions", "Session1", "Selected_rules.xlsx") \%>\%
  import_data() \%>\%
  simplify_ruleset(target_vec = Target, target_data = SpecificDTM) \%>\%
  pull(rule)
}
}
