% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting.R
\name{source_session_summary_to_list}
\alias{source_session_summary_to_list}
\title{Format records' source distribution as a list}
\usage{
source_session_summary_to_list(source_summary)
}
\arguments{
\item{source_summary}{A data frame produced by
\code{\link{summarise_sources_by_session}()}.}
}
\value{
A hierarchical list with record distribution by source and session.
}
\description{
Takes the output of \code{\link{summarise_sources_by_session}()} and formats it
as a list.
}
\examples{

\dontrun{
# Describe source distribution by source and in total
source_summary <- summarise_sources_by_session(add_global_totals = TRUE)

# Transform it into a list
source_session_summary_to_list(source_summary)
}
}
