% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting.R
\name{summarise_annotations}
\alias{summarise_annotations}
\title{Describe results of a Classification/Review session}
\usage{
summarise_annotations(
  session_name,
  sessions_folder = getOption("baysren.sessions_folder"),
  remove_empty_columns = TRUE,
  remove_raw_data = TRUE
)
}
\arguments{
\item{session_name}{A session identifier corresponding to folders into the
\code{sessions_folder} folder.}

\item{sessions_folder}{Where to find the \code{sessions} folders.}

\item{remove_empty_columns}{If \code{TRUE}, shows change columns even if no
changes of that type ever happened.}

\item{remove_raw_data}{Remove the \code{tot_reviewed_} and
\code{total_records_} columns from the output, which store data and column
name in machine readable column formats.}
}
\value{
A data frame reporting for a session the number of positive and
negative matches after each Classification/Review iteration and the
specific changes corresponding to each iteration.
}
\description{
Take a session identifier as input and describe the changes in the number of
positive and negative matches after each Classification/Review iteration. It
is necessary to run \code{\link{consolidate_results}()} before this command,
otherwise the results would not consider the changes made through the manual
review of the automatic classification.
}
\examples{

}
