% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting.R
\name{summarise_annotations_by_session}
\alias{summarise_annotations_by_session}
\title{Describe results of all Classification/Review sessions}
\usage{
summarise_annotations_by_session(
  sessions_folder = getOption("baysren.sessions_folder"),
  remove_empty_columns = TRUE,
  remove_raw_data = TRUE
)
}
\arguments{
\item{sessions_folder}{A repository where session folders are stored.}

\item{remove_empty_columns}{If \code{TRUE}, shows change columns even if no
changes of that type ever happened.}

\item{remove_raw_data}{Remove the \code{tot_reviewed_} and
\code{total_records_} columns from the output, which store data and column
name in machine readable column formats.}
}
\value{
A data frame reporting for each session the number of positive and
negative matches after each Classification/Review iteration and the
specific changes corresponding to each iteration.
}
\description{
Applies \code{\link{summarise_annotations}()} to all session in the
\code{sessions_folder} folder.
}
\examples{

\dontrun{

summarise_annotations_by_session()
}
}
