% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting.R
\name{summarise_by_source}
\alias{summarise_by_source}
\title{Record distribution between sources in an Annotation file}
\usage{
summarise_by_source(
  annotation_file,
  as_data_frame = FALSE,
  add_session_totals = TRUE
)
}
\arguments{
\item{annotation_file}{An annotation data frame or a file path to it.}

\item{as_data_frame}{Whether to return the results as a data frame
(\code{TRUE}) or a list (\code{FALSE}).}

\item{add_session_totals}{Whether to add the total number of record.}
}
\value{
For each source:\item{Records}{The number of records related to the
source.}\item{\\% over total}{the percentage over the total
records.}\item{Source specific records}{the number of record derived only
from that source.}\item{\\% over source total}{the value in \code{Source
  specific records} as percentage over the source total.} If
\code{as_data_frame} is \code{TRUE}, the results are returned as a data
frame with a row for each source, otherwise as a nested list.
\code{add_session_totals} adds another entry with the total number of
records in the \code{annotation_file}.
}
\description{
Summarises the distribution of the source databases among the citation
records in an Annotation data set. It also reports the fraction of records
that are unique for each database.
}
\examples{

\dontrun{

records <- get_session_files('Session1')$Records

summarise_by_source(records)

}
}
