% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reporting.R
\name{summarise_sources_by_session}
\alias{summarise_sources_by_session}
\title{Record distribution between sources for each session}
\usage{
summarise_sources_by_session(
  sessions = list.files(sessions_folder),
  sessions_folder = getOption("baysren.sessions_folder"),
  add_global_totals = TRUE,
  keep_session_label = FALSE,
  ...
)
}
\arguments{
\item{sessions}{A vector of session identifiers corresponding to folders into
the \code{sessions_folder} folder.}

\item{sessions_folder}{Where to find the \code{sessions} folders.}

\item{add_global_totals}{Add results for all sessions considered together.}

\item{keep_session_label}{Add a column which groups rows by session. Useful
for subsequent sub-setting of the results.}

\item{...}{Other arguments passed to \code{\link[=summarise_by_source]{summarise_by_source()}}.}
}
\value{
A data frame with the number and fraction of total records per source
and number and fraction of source-specific records, grouped by session. An
extra group with the overall results is reported if
\code{add_global_totals} is \code{TRUE}.
}
\description{
Applies \code{\link[=summarise_by_source]{summarise_by_source()}} to all selected sessions.
}
\examples{
\dontrun{
# Describe source distribution by source and in total
summarise_sources_by_session(add_global_totals = TRUE)
}
}
