% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Helpers.R
\name{today}
\alias{today}
\title{Override lubridate:today() which always complain about the missing
timezone}
\usage{
today()
}
\description{
Override lubridate:today() which always complain about the missing
timezone
}
