% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{tokenize_authors}
\alias{tokenize_authors}
\title{Tokenize authors in records}
\usage{
tokenize_authors(corpus)
}
\arguments{
\item{corpus}{A vector of author fields from an annotation data set.}
}
\value{
The tokenized author list.
}
\description{
Use different tokenization approaches based on author field format.
}
