% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{tokenize_keywords}
\alias{tokenize_keywords}
\title{Tokenize keywords in records}
\usage{
tokenize_keywords(keywords)
}
\arguments{
\item{corpus}{A vector of keywords fields from an annotation data set.}
}
\value{
The tokenized keyword list.
}
\description{
Clean up the keyword fields in the records.
}
