% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{tokenize_MESH}
\alias{tokenize_MESH}
\title{Tokenize MESH keywords in records}
\usage{
tokenize_MESH(mesh)
}
\arguments{
\item{corpus}{A vector of MESH keywords fields from an annotation data set.}
}
\value{
The tokenized MESH keyword list.
}
\description{
Clean up the keyword fields in the records.
}
