% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NLP.R
\name{tokenize_text}
\alias{tokenize_text}
\title{Clean up text into tokens}
\usage{
tokenize_text(corpus)
}
\arguments{
\item{corpus}{A vector of text documents.}
}
\value{
A cleaned-up vector of text documents.
}
\description{
Removes stop words, punctuation, auxiliary verbs. Lemmatizes text. Changes to
lower-case.
}
